--- src/pages/[uid].js
@@ -1,48 +1,0 @@
-import Head from "next/head";
-import { SliceZone } from "@prismicio/react";
-import * as prismic from "@prismicio/client";
-
-import { createClient } from "@/prismicio";
-import { components } from "@/slices";
-import { Layout } from "@/components/Layout";
-
-export default function Page({ page, navigation, settings }) {
-  return (
-    <Layout navigation={navigation} settings={settings}>
-      <Head>
-        <title>
-          {prismic.asText(page.data.title)} |{" "}
-          {prismic.asText(settings.data.name)}
-        </title>
-      </Head>
-      <SliceZone slices={page.data.slices} components={components} />
-    </Layout>
-  );
-}
-
-export async function getStaticProps({ params, previewData }) {
-  const client = createClient({ previewData });
-
-  const page = await client.getByUID("page", params.uid);
-  const navigation = await client.getSingle("navigation");
-  const settings = await client.getSingle("settings");
-
-  return {
-    props: {
-      page,
-      navigation,
-      settings,
-    },
-  };
-}
-
-export async function getStaticPaths() {
-  const client = createClient();
-
-  const pages = await client.getAllByType("page");
-
-  return {
-    paths: pages.map((page) => prismic.asLink(page)),
-    fallback: false,
-  };
-}
